refactor: extract useAccount from useWallet
